requirements.txt
## General
python-dotenv

## Web crawling and scraping
requests
requests-cache
# requests-html
pygooglenews==0.1.3
googlenewsdecoder
beautifulsoup4==4.12.2
crawl4ai

## Text & URL Parsing
feedparser
urllib3
# lxml
# html5lib # https://stackoverflow.com/a/73403774

## Data Analysis
pandas
numpy

## Database
flask-sqlalchemy
flask-migrate

## Machine Learning
openai
llama-index
llama-index-core
# langchain
# langchain-core
# langchain-community

## Vector DB
chromadb
# faiss-cpu

## Web server
flask
# flask_cors
# flask[async]

## Deployment
gunicorn

## Miscellaneous
ipykernel
